requirements.txt
flask
pika
pyyaml
